// apps/legacy/src/app/[roomid]/map/types.ts
export interface Point {
    x: number;
    y: number;
}
// ...
export interface MusicZone {
    id: string;
    x: number;
    y: number;
    radius: number;
    trackId?: string;
    volume: number;
    name?: string;
    color?: string;
    url?: string | null;
    cityId?: string | null;
}